Creation of a Singleton exceptionHandlerController
Move the exception close to the class where they are throw
make ressourceCreationError an internal class of the exception handler
