+++ api/src/main/java/com/michelin/ns4kafka/controllers/ResourceValidationException.java
@@ -1,4 +1,4 @@
-package com.michelin.ns4kafka.exception;
+package com.michelin.ns4kafka.controllers;
 
 import lombok.Getter;
 
@@ -7,7 +7,7 @@ import java.util.List;
 public class ResourceValidationException extends RuntimeException {
     private static final long serialVersionUID = 32400191899153204L;
     @Getter
-    List<String> validationErrors;
+    private final List<String> validationErrors;
 
     public ResourceValidationException(List<String> validationsErrors) {
         this.validationErrors = validationsErrors;
