+++ api/src/main/java/com/michelin/ns4kafka/validation/TopicValidator.java
@@ -2,7 +2,6 @@ package com.michelin.ns4kafka.validation;
 
 import com.michelin.ns4kafka.models.Namespace;
 import com.michelin.ns4kafka.models.Topic;
-import com.michelin.ns4kafka.exception.FieldValidationException;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
